Convert APIServerPort to uint16
This is a TCP port, which is an unsigned 16 bit integer. We were
previously allowing it to be any integer.

This is a breaking change, but as any working configuration would have
to have a value which passes the new validation it should not break any
working configuration.

+++ api/v1beta1/openstackcluster_types.go
@@ -133,9 +133,9 @@ type OpenStackClusterSpec struct {
 	APIServerFixedIP optional.String `json:"apiServerFixedIP,omitempty"`
 
 	// APIServerPort is the port on which the listener on the APIServer
-	// will be created
+	// will be created. If specified, it must be an integer between 0 and 65535.
 	// +optional
-	APIServerPort optional.Int `json:"apiServerPort,omitempty"`
+	APIServerPort optional.UInt16 `json:"apiServerPort,omitempty"`
 
 	// ManagedSecurityGroups determines whether OpenStack security groups for the cluster
 	// will be managed by the OpenStack provider or whether pre-existing security groups will
